Don't use telescope's git_files to find files anymore
It used to be that the `git_files` method would fall through if you
weren't in a git repo. It doesn't do that any longer. It also doesn't
find files that are untracked, which has come up a lot for me lately.

I can probably customize it but I'd rather default back to the normal
finder for now so I can take my time re-learning how it works.