Streamline my LSP config a bit
Took lots of inspiration from `@franciscoj` here and moved most of the
stuff in the lsp handlers module a new config module. The server
configuration has also been changed to use this new config module. Most
of the code in the handlers module isn't about handlers now and could
probably use a rename. :)