feat(neovim): Make it easier to work with highlight groups
The new module is called `mattr-.colors` and, so far, only has a method
to get a slimmed down highlight group with the foreground only. This is
used to add a bit of spice for the command log in the status line and
the mode in the status line (i.e. `recording @a`)