feat(neovim): hoist the util class to a global
The various `require("mattr-.util.whatever")` calls were starting to bug
me but I didn't have a better way to handle these. I saw this cool hack
while reading the LazyVim codebase where a global is added pointing to
the root of the util module and a metatable in that module handles
looking up things from either lazy.nvim's util library or our local util
library. It was so cool that I decided to steal it.