feat(neovim): add custom notification wrappers
`vim.notify` is great! let's wrap it in some custom methods in the ui
module to add some defaults.

`vim.notify` is backed by the `rcarriga/nvim-notify` in my config and
nicely passes any extra options on to the backend, which is what makes
this work so well.

Next up: Replace instances of lazy.nvim's helpers with these ones