refactor(neovim): replace the features module
The features system was very cool code but ultimately, it was way too
complex for what I was using it for. Replace it with a `tools` module in
the `util` namespace.

The server configurations for both sorbet and rubocop are also ported
over to the new `tools` module and become much simpler since we're no
longer doing any command overriding and letting mason handle most
things.