fix(macOS): general Nix brokenness
I think I was a bit too opinionated here and didn't really know or
understand that Nix on macOS would not be capable of being configured
the same way it is in NixOS. In hindsight, it's obvious, but it wasn't
when I first set this up.

Instead, let's tone down how we aggressive we are about configuring Nix
on macOS to see if we get a better out of the box experience.