feat(neovim): add slimmed down toggle library and use it
This imports the toggle library from folke/snacks.nvim and modifies it
to use my own notification ui and removes some of the things I don't
have toggles set up for yet.

As part of this, we're also setting up a custom toggle for mini.pairs to
turn it on and off as a demo.